backend/requirements.txt
fastapi==0.115.12
pydantic[email]==2.8.2
uvicorn==0.22.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
passlib[argon2]==1.7.4
python-jose==3.3.0
psycopg2-binary==2.9.9
python-multipart==0.0.9
fastapi_pagination==0.12.24
trio==0.25.1
websockets==12.0
aiohttp==3.9.5
minio==7.1.12
pillow==10.4.0
typing-extensions==4.12.2
colorama==0.4.6
aio_pika==9.5.4
redis==5.2.1
apscheduler==3.10.4
md2tgmd==0.3.9
prometheus_fastapi_instrumentator==7.1.0
python-dateutil==2.9.0.post0